refactor(lsp): move WorkDoneProgressHelper to lsp.client to break lsp↔context cycle
Used by context indexing (ServerContext/ServerContextProvider) for progress
reporting, the helper depends only on the lsp.client holders. Placing it in the
lower lsp.client layer removes the context→lsp back-edge.
