+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/lsp/client/WorkDoneProgressHelper.java
@@ -19,11 +19,7 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with BSL Language Server.
  */
-package com.github._1c_syntax.bsl.languageserver.lsp;
-
-
-import com.github._1c_syntax.bsl.languageserver.lsp.client.ClientCapabilitiesHolder;
-import com.github._1c_syntax.bsl.languageserver.lsp.client.LanguageClientHolder;
+package com.github._1c_syntax.bsl.languageserver.lsp.client;
 
 import lombok.AllArgsConstructor;
 import lombok.RequiredArgsConstructor;
